pyproject.toml
[tool.poetry]
name = "lisa-ai"
version = "0.1.0"
description = "AI-assisted system for psychoeducational report generation"
authors = ["Lisa AI Team"]

[tool.poetry.dependencies]
python = "^3.10"
crewai = "^0.1.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
streamlit = "^1.24.0"
python-docx = "^0.8.11"
PyPDF2 = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 